feat(ci): extract analytics and user feedback api keys into env variables

+++ desktop/domains/feedbackWidget/index.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React, { useContext, useEffect, useState } from "react";
 
+import { USERSNAP_GLOBAL_API_KEY } from "@aca/desktop/lib/env";
 export const UsersnapContext = React.createContext<any>(null);
 
 export const UsersnapProvider = ({
@@ -13,23 +14,25 @@ export const UsersnapProvider = ({
   const [usersnapApi, setUsersnapApi] = useState(null);
 
   useEffect(() => {
-    let usersnapApi: any = null;
-    window.onUsersnapCXLoad = function (api) {
-      api.init(initParams);
-      usersnapApi = api;
-      setUsersnapApi(api);
-    };
-    const script = document.createElement("script");
-    script.defer = true;
-    script.src = "https://widget.usersnap.com/global/load/ac75bf8d-d645-450c-9f62-42054ff9458f?onload=onUsersnapCXLoad";
-    document.head.appendChild(script);
+    if (USERSNAP_GLOBAL_API_KEY) {
+      let usersnapApi: any = null;
+      window.onUsersnapCXLoad = function (api) {
+        api.init(initParams);
+        usersnapApi = api;
+        setUsersnapApi(api);
+      };
+      const script = document.createElement("script");
+      script.defer = true;
+      script.src = `https://widget.usersnap.com/global/load/${USERSNAP_GLOBAL_API_KEY}?onload=onUsersnapCXLoad`;
+      document.head.appendChild(script);
 
-    return () => {
-      if (usersnapApi) {
-        usersnapApi.destroy();
-      }
-      script.remove();
-    };
+      return () => {
+        if (usersnapApi) {
+          usersnapApi.destroy();
+        }
+        script.remove();
+      };
+    }
   }, []);
 
   return <UsersnapContext.Provider value={usersnapApi}>{children}</UsersnapContext.Provider>;
